functions/input_handler/requirements.txt
google-cloud-storage==2.7.0
google-cloud-pubsub==2.13.0
functions-framework==3.3.0